SASS-6481: Load prior data for Income journey (#136)
It loads the submitted answers. The submitted answers are also called 'journey answers'. There is a small logic for merging the submitted answers with the user session (called user answers).

The logic is only triggered on the CYA page and only when there is no data for the journey in session.

The logic is:
    no session data + no submitted answers == no data
    no session data + submitted answers == submitted answers
    session data for journey A + submitted answers for journey B == answers merged A + B
    session data for journey A + submitted answers A == session data for journey A

Also few improvements for everyday development:
    I removed wart remover, as it caused more problems than helped
    You can now add 'PLAY_ENV=Dev' to your env /intellij env variables, refresh SBT and then you'll have no warn failure checks in intellij.