Add NICS sample event (first version) (#405)
I had to add submit our not-ready class 2 NICs to backend to emit the event, it causes a failure right now because we don't have a backend. It will be fixed in the next PR.

It is an example event so PA can start thinking what else we need. There will be a formal process which would take couple of sprints to establish final version of those evens.